contact_extractor/requirements.txt
imaplib2>=3.5
PyYAML>=6.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.2
phonenumbers 
spacy
#-m spacy download en_core_web_sm